+++ routes/routerExercise.ts
@@ -9,32 +9,19 @@ export const exerciseRouter = Router();
 exerciseRouter.post("/", async (request, response) => {
     //req.body
     const exercise: ExerciseInterface = request.body;
-
-    try {
-        const saveExercise = await Exercise.create(exercise);
-
-        return response.status(201).json({
-            savedID: saveExercise.id,
-            message: "Exercício inserido no sistema"
-        });
-    } catch (error) {
-        return response.status(500).json({ error: error });
-    }
-});
-
-exerciseRouter.get("/", async (request, response) => {
-
     const token = await verifyToken(request.headers.authorization)
 
     if (token) {
         if ((token as UserInterface).role == "admin") {
             try {
+                const saveExercise = await Exercise.create(exercise);
 
-                const exercises = await Exercise.find()
-                return response.status(200).json(exercises)
-
+                return response.status(201).json({
+                    savedID: saveExercise.id,
+                    message: "Exercício inserido no sistema"
+                });
             } catch (error) {
-                return response.status(500).json({ error: error })
+                return response.status(500).json({ error: error });
             }
         } else {
             return response.status(403).json({ message: "Você não possui este acesso" })
@@ -44,21 +31,46 @@ exerciseRouter.get("/", async (request, response) => {
     }
 });
 
+exerciseRouter.get("/", async (request, response) => {
+
+    const token = await verifyToken(request.headers.authorization)
+
+    if (token) {
+
+        try {
+
+            const exercises = await Exercise.find()
+            return response.status(200).json(exercises)
+
+        } catch (error) {
+            return response.status(500).json({ error: error })
+        }
+
+    } else {
+        return response.status(401).json({ message: "Token Inválido" })
+    }
+});
+
 exerciseRouter.get("/:id", async (request, response) => {
     const id = request.params.id;
+    const token = await verifyToken(request.headers.authorization)
 
-    try {
-        // findONe({ _id: id})
-        const exercise = await Exercise.findById(id);
 
-        if (!exercise) {
-            return response
-                .status(422)
-                .json({ message: "O exercício não foi encontrado" });
+    if (token) {
+        try {
+            const exercise = await Exercise.findById(id);
+
+            if (!exercise) {
+                return response
+                    .status(422)
+                    .json({ message: "O exercício não foi encontrado" });
+            }
+            return response.status(200).json(exercise);
+        } catch (error) {
+            return response.status(500).json({ error: error });
         }
-        return response.status(200).json(exercise);
-    } catch (error) {
-        return response.status(500).json({ error: error });
+    } else {
+        return response.status(401).json({ message: "Token Inválido" })
     }
 });
 
@@ -66,21 +78,29 @@ exerciseRouter.get("/:id", async (request, response) => {
 
 exerciseRouter.patch("/:id", async (request, response) => {
     const id = request.params.id; // se alterar em cima altera o parâmetro
-
+    const token = await verifyToken(request.headers.authorization)
     const exercise: ExerciseInterface = request.body;
 
-    try {
-        await Exercise.findByIdAndUpdate(id, exercise);
+    if (token) {
+        if ((token as UserInterface).role == "admin") {
+            try {
+                await Exercise.findByIdAndUpdate(id, exercise);
 
-        return response.status(200).json(exercise);
-    } catch (error) {
-        return response.status(500).json({ error: error });
+                return response.status(200).json(exercise);
+            } catch (error) {
+                return response.status(500).json({ error: error });
+            }
+        } else {
+            return response.status(403).json({ message: "Você não possui este acesso" })
+        }
+    } else {
+        return response.status(401).json({ message: "Token Inválido" })
     }
 });
 
 exerciseRouter.delete("/:id", async (request, response) => {
     const id = request.params.id;
-
+    const token = await verifyToken(request.headers.authorization)
     const exercise = await Exercise.findById(id);
 
     if (!exercise) {
@@ -89,11 +109,19 @@ exerciseRouter.delete("/:id", async (request, response) => {
             .json({ message: "O exercício não foi encontrado" });
     }
 
-    try {
-        await Exercise.findByIdAndDelete(id);
+    if (token) {
+        if ((token as UserInterface).role == "admin") {
+            try {
+                await Exercise.findByIdAndDelete(id);
 
-        return response.status(200).json({ message: "Exercício deletado" });
-    } catch (error) {
-        return response.status(500).json({ error: error });
+                return response.status(200).json({ message: "Exercício deletado" });
+            } catch (error) {
+                return response.status(500).json({ error: error });
+            }
+        } else {
+            return response.status(403).json({ message: "Você não possui este acesso" })
+        }
+    } else {
+        return response.status(401).json({ message: "Token Inválido" })
     }
 });
